refactor: [BEdita/API] replace deprecated methods with new ones
