+++ plugins/BEdita/API/tests/IntegrationTest/SortQueryStringTest.php
@@ -13,7 +13,6 @@
 namespace BEdita\API\Test\IntegrationTest;
 
 use BEdita\API\TestSuite\IntegrationTestCase;
-use Cake\Network\Exception\BadRequestException;
 use Cake\Utility\Hash;
 
 /**
