+++ plugins/BEdita/API/tests/TestCase/Error/ExceptionRendererTest.php
@@ -17,9 +17,9 @@ use BEdita\API\Error\ExceptionRenderer;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Event\Event;
+use Cake\Http\Response;
 use Cake\Network\Exception\InternalErrorException;
 use Cake\Network\Exception\NotFoundException;
-use Cake\Network\Response;
 use Cake\TestSuite\TestCase;
 
 /**
@@ -131,7 +131,7 @@ class ExceptionRendererTest extends TestCase
     public function testErrorDetails($errorMessage, $title, $detail = '', $code = '')
     {
         $renderer = new ExceptionRenderer(new NotFoundException($errorMessage));
-        $renderer->controller->request->env('HTTP_ACCEPT', 'application/json');
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', 'application/json');
         $response = $renderer->render();
 
         $responseBody = json_decode((string)$response->getBody(), true);
@@ -215,7 +215,7 @@ class ExceptionRendererTest extends TestCase
     {
         Configure::write($config);
         $renderer = new ExceptionRenderer(new NotFoundException());
-        $renderer->controller->request->env('HTTP_ACCEPT', $accept);
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', $accept);
 
         $this->assertEquals($expected, $renderer->isHtmlToSend());
     }
@@ -280,7 +280,7 @@ class ExceptionRendererTest extends TestCase
         }
 
         $renderer = new ExceptionRenderer(new NotFoundException('test html'));
-        $renderer->controller->request->env('HTTP_ACCEPT', 'text/html');
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', 'text/html');
         $response = $renderer->render();
 
         $this->checkResponseHtml($renderer, $response, $config['debug']);
@@ -308,9 +308,9 @@ class ExceptionRendererTest extends TestCase
         }
 
         $renderer = new ExceptionRenderer(new NotFoundException('test html'));
-        $renderer->controller->request->env('HTTP_ACCEPT', 'text/html');
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', 'text/html');
 
-        $renderer->controller->eventManager()->on('Controller.beforeRender', function (Event $event) {
+        $renderer->controller->getEventManager()->on('Controller.beforeRender', function (Event $event) {
             // force missing layout exception
             $event->getSubject()->viewBuilder()->setLayoutPath('find_me_if_you_can');
         });
@@ -325,7 +325,7 @@ class ExceptionRendererTest extends TestCase
      *  Perform some asserts to check html response
      *
      * @param \BEdita\API\Error\ExceptionRenderer $renderer
-     * @param \Cake\Network\Response $response
+     * @param \Cake\Http\Response $response
      * @param int $debug
      * @return void
      */
@@ -427,7 +427,7 @@ class ExceptionRendererTest extends TestCase
         }
 
         $renderer = new ExceptionRenderer(new NotFoundException('test html'));
-        $renderer->controller->request->env('HTTP_ACCEPT', $accept);
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', $accept);
         $response = $renderer->render();
 
         $this->checkResponseJson($renderer, $response, $config['debug']);
@@ -455,9 +455,9 @@ class ExceptionRendererTest extends TestCase
         }
 
         $renderer = new ExceptionRenderer(new NotFoundException('test html'));
-        $renderer->controller->request->env('HTTP_ACCEPT', $accept);
+        $renderer->controller->request = $renderer->controller->request->withEnv('HTTP_ACCEPT', $accept);
 
-        $renderer->controller->eventManager()->on('Controller.beforeRender', function () {
+        $renderer->controller->getEventManager()->on('Controller.beforeRender', function () {
             throw new InternalErrorException();
         });
 
@@ -470,7 +470,7 @@ class ExceptionRendererTest extends TestCase
      *  Perform some asserts to check JSON response
      *
      * @param \BEdita\API\Error\ExceptionRenderer $renderer
-     * @param \Cake\Network\Response $response
+     * @param \Cake\Http\Response $response
      * @param int $debug
      * @return void
      */
